:truck: chore: move Footer, Newsletter, and Products files to their respective component folders
