+++ src/App.jsx
@@ -1,9 +1,9 @@
 import './App.css'
 import { Header } from './components/Header/Header.jsx'
 import { MainContent } from './components/MainContent/MainContent.jsx'
-import { Products } from './components/Products.jsx'
-import { Newsletter } from './components/Newsletter.jsx'
-import { Footer } from './components/Footer.jsx'
+import { Products } from './components/Products/Products.jsx'
+import { Newsletter } from './components/Newsletter/Newsletter.jsx'
+import { Footer } from './components/Footer/Footer.jsx'
 
 function App() {
     return <div>
